+++ src/forms/index.js
@@ -1,4 +1,4 @@
 const { form, Form } = require('./form');
-const { field, FieldDesriptor } = require('../services/fields');
+const { field, FieldDesriptor } = require('./field');
 
 module.exports = { form, Form, field, FieldDesriptor };
